+++ refresh_tokens/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class RefreshTokensConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "project_name.refresh_tokens"
+    name = "refresh_tokens"
