+++ requirements_dev.txt
@@ -2,7 +2,7 @@
 Django==5.1.3
 django-environ==0.11.2
 gunicorn==23.0.0
-psycopg2==3.2.3
+psycopg2-binary==2.9.10
 
 # Rest apis
 djangorestframework==3.15.2
@@ -10,6 +10,15 @@ djangorestframework_simplejwt==5.3.1
 django-cors-headers==4.6.0
 drf-yasg==1.21.8
 whitenoise==6.8.2
+
+# Code quality
+# ------------------------------------------------------------------------------
+flake8==7.1.1
+flake8-isort==6.1.1
+coverage==7.6.8
+black==24.10.0
+pylint-django==2.6.1
+pre-commit==4.0.1
 
 # Certificates
 cryptography==43.0.3
